pyproject.toml
[project]
name = "dana-cockfight-bot"
version = "0.1.0"
description = "Telegram bot for generating Dana cockfight images using AI"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=12.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

